requirements.txt
pytorch-lightning==1.1.2
torch-summary==1.4.3
numpy==1.20.1
opencv==4.5.0
